Student Productivity Planner
This project plans on creating a planner that organizes and optimizes the time management of a student. Features include a to-do list, calender, and organizer. Aside from that, the unique feature of this program compared to existing planners is that they will get the information, profs, assignments of the student after being inputted in order to follow up tasks and suggest the best times to start the project given the several deadlines and tasks in their to-do list. This program can also link to several word processor and educational apps for the student to be on track on all their pending and done outputs.